requirements.txt
discord.py>=2.3.0
asyncpg>=0.28.0
cryptography>=41.0.0
python-dotenv>=1.0.0
yt-dlp>=2023.0.0
mutagen>=1.46.0
requests>=2.31.0
aiohttp>=3.8.0